// apps/frontend/core-dapp/src/types/market.types.ts
export enum HistoricalDataQueryType {
  HOUR = 'HOUR',
  DAY = 'DAY',
}

export enum ChartView {
  VALUATION = 'Valuation',
  VOLUME = 'Volume',
  LIQUIDITY = 'Liquidity',
}

export enum ChartTimeframe {
  DAY = '1D',
  WEEK = '1W',
  MONTH = '1M',
  YEAR = '1Y',
  ALL = 'ALL',
}

export type Range = [number, number]
// ...
export type MarketHistoryData = {
  timestamp: number
  volume: number
  valuation: number
  liquidity?: number
}

export type SeparatedMarketHistoryData = {
  long: MarketHistoryData[]
  short: MarketHistoryData[]
}
// ...
export type PoolDayData = Pick<
  PoolDayDataModelType,
  | 'date'
  | 'liquidity'
  | 'token0Price'
  | 'token1Price'
  | 'id'
  | 'sqrtPrice'
  | 'volumeToken0'
  | 'volumeToken1'
> &
  PoolTokensId

export type PoolDayDatas = PoolDayData[]

export type PoolsDayDatas = {
  longTokenPool?: PoolDayDatas
  shortTokenPool?: PoolDayDatas
}

export type PoolHourData = Pick<
  PoolHourDataModelType,
  | 'id'
  | 'liquidity'
  | 'periodStartUnix'
  | 'sqrtPrice'
  | 'token0Price'
  | 'token1Price'
  | 'volumeToken0'
  | 'volumeToken1'
> &
  PoolTokensId

export type PoolHourDatas = PoolHourData[]

export type PoolsHourDatas = {
  longTokenPool?: PoolHourDatas
  shortTokenPool?: PoolHourDatas
}

export type MarketHistoryDataMap = {
  [timestamp: number]: MarketHistoryData
}

// market utils types

export type FormatPoolsHistoricalDatasOptions = {
  payoutRange: Range
  tokenAddresses: {
    long: string
    short: string
  }
  valuationRange: Range
  endTimeInMs?: number
  startTimeInMs?: number
  type: HistoricalDataQueryType
}